pyproject.toml
[project]
name = "Multimodal_RAG"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Gopi Pitchai",email = "gopipitchai@gmail.com"},
]
dependencies = ["docling>=2.15.0", "fastapi>=0.115.6", "uvicorn>=0.34.0", "langchain>=0.3.14", "sqlalchemy>=2.0.36", "langchain-community>=0.3.14", "langchain-text-splitters>=0.3.5", "langchain-huggingface>=0.1.2", "langchain-openai>=0.2.14", "langchain-chroma>=0.2.0", "langchain-core>=0.3.29", "langchain-experimental>=0.3.4", "pillow>=10.4.0", "langchain-ollama>=0.2.2", "pandas>=2.2.3", "gpt4all>=2.8.2", "open-clip-torch>=2.30.0", "python-multipart>=0.0.20", "requests>=2.32.3", "streamlit>=1.41.1", "unstructured[all-docs]>=0.16.12", "ipykernel>=6.29.5", "optimum>=1.23.3", "faiss-cpu>=1.9.0.post1", "langgraph>=0.2.62", "langchain-nomic>=0.1.4", "lm-format-enforcer>=0.10.9", "crewai>=0.95.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
